[Enhancement] Create pull request template (#86)
**Issue:** #85

**Description:**
Add pull request template to the repository.
The purpose of this template is to unify pull request titles and
descriptions, to make sure, that all the needed informations are
contained in commit message after squashing and merging of the PR (it is assumed, that PR title/description will be used as final commit message). This should make browsing of the repository and it's history easier, and more intuitive. Additionally it should make reviewing of the PR easier, as the reviewer should know the context of the changes after reading PR description.