[CMMD] Create a pipeline for CMMD dataset  (#91)
**Issue:** #80

**Description:**
Creates a preprocessing pipeline for images stored in CMMD dataset.
The purpose is to add the ability to preprocess another medical image
dataset to the repository. The dataset is CMMD (The Chinese Mammography
Database) containing 5,202 MG (Mammography) breast images.

Additionally the file steps/convert_dcm2png.py is changed. Previously,
for this step to work correctly, the masks_path argument had to be
provided in the config/runner_config.py file, even if the dataset did
not have masks. Now the masks_path argument is optional.

---------

Co-authored-by: Kacper Rogala <kacper.rogala6@wp.pl>